\name{EasyqpcR-package}
\alias{EasyqpcR-package}
\alias{EasyqpcR}
\docType{package}
\title{
EasyqpcR for low-throughput real-time quantitative PCR data
        analysis

}
\description{
This package is based on the qBase algorithms published by 
 Hellemans et al. in 2007. The EasyqpcR package allows you to import easily qPCR
 data files as described in the vignette. Thereafter, you can calculate 
 amplification efficiencies, relative quantities and their standard errors, 
 normalization factors based on the best reference genes choosen (using the 
 SLqPCR package), and then the normalized relative quantities, the NRQs scaled 
 to your control and their standard errors. This package has been created for 
 low-throughput qPCR data analysis.
}
\details{
\tabular{ll}{
Package: \tab EasyqpcR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-31\cr
License: \tab GPL (>=2)\cr
}

}
\author{
Sylvain Le Pape

Maintainer: Sylvain Le Pape <sylvain.le.pape@univ-poitiers.fr>

}
\references{
Jan Hellemans, Geert Mortier, Anne De Paepe, Frank Speleman and Jo Vandesompele.
 qBase relative quantification framework and software for management and
 automated analysis of real-time quantitative PCR data. Genome Biology 2007,
 8:R19 (doi:10.1186/gb-2007-8-2-r19). <url:http://genomebiology.com/2007/8/2/R19>
}
\keyword{qPCR analysis}

