\name{EasyqpcR-package}
\alias{EasyqpcR-package}
\alias{EasyqpcR}
\docType{package}
\title{
Functions to analyse real-time quantitative PCR data at IRTOMIT-INSERM U1082
}
\description{
This package contains functions to analyse real-time quantitative PCR data at
 IRTOMIT-INSERM U1082. The algorithm used is the one published by Hellemans et
 al. (2007). It permits the calculation of the normalization and calibration
 factors, the relative quantities, the normalized relative quantities which can
 then be scaled to your control group.
}
\details{
\tabular{ll}{
Package: \tab EasyqpcR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-31\cr
License: \tab GPL (>=2)\cr
}

}
\author{
Sylvain Le Pape

Maintainer: Sylvain Le Pape <sylvain.le.pape@univ-poitiers.fr>

}
\references{
Jan Hellemans, Geert Mortier, Anne De Paepe, Frank Speleman and Jo Vandesompele.
 qBase relative quantification framework and software for management and
 automated analysis of real-time quantitative PCR data. Genome Biology 2007,
 8:R19 (doi:10.1186/gb-2007-8-2-r19). <url:http://genomebiology.com/2007/8/2/R19>
}
\keyword{qPCR analysis}

